Added to stretch goal
- Since I realized we can use other technologies
for stretch goals because another student
used Bootstrap. I implemented some PHP
into this project for the services.php page.

- Created partials folders to hold the footer
And header, making for less redundant
html markup across various pages

- Created an array in PHP to hold multiple
header image file paths. This way when
visiting the services.php page a function
is used to randomize the order of the array
so a different picture is used every time the
page is loaded.